# feat/rankings: optimize Steam ID comparison