--- a/rankings/prefetch.go
+++ b/rankings/prefetch.go
@@ -25,7 +25,7 @@ func readRecords() []Record {
 	return records
 }
 
-func readOverrides() map[string]map[string]int {
+func readOverrides() map[SteamID]map[string]int {
 	overridesFile, err := os.Open("./input/overrides.json")
 	if err != nil {
 		log.Fatalln(err.Error())
@@ -35,7 +35,7 @@ func readOverrides() map[string]map[string]int {
 	if err != nil {
 		log.Fatalln(err.Error())
 	}
-	overrides := map[string]map[string]int{}
+	overrides := map[SteamID]map[string]int{}
 	err = json.Unmarshal(overridesFileBytes, &overrides)
 	if err != nil {
 		log.Fatalln(err.Error())