Since the Steam ID is a 64bit integer we should compare it in registers directly instead of comparing a string byte by byte. The string representation is only needed when constructing the URL to call the profile API and when exporting the data.

It's a µ-optimization so this does not speed up the program too much since the biggest slowdown will always be network requests.

Profiled this small code section:
```go
start := time.Now()

for _, profile := range data.Response.Players {
	for _, player := range players {
		if player.SteamID == profile.SteamID {
			player.AvatarLink = profile.AvatarFull
			player.Username = profile.PersonaName
		}
	}
}

duration := time.Since(start)
fmt.Printf("took: %v\n", duration)
```

```bash
# main
2024/11/16 22:01:23 fetching info for 100 players
took: 262.425µs
2024/11/16 22:01:24 fetching info for 100 players
took: 246.73µs
2024/11/16 22:01:25 fetching info for 92 players
took: 209.21µs

# this pr ~8x faster
2024/11/16 22:05:00 fetching info for 100 players
took: 31.959µs
2024/11/16 22:05:00 fetching info for 100 players
took: 31.688µs
2024/11/16 22:05:01 fetching info for 92 players
took: 26.985µs
```